+++ validationframework-core/src/main/java/com/google/code/validationframework/api/property/WritableListProperty.java
@@ -79,7 +79,7 @@ public interface WritableListProperty<W> {
     /**
      * @see java.util.List#retainAll(Collection)
      */
-    boolean retainAll(Collection<?> c);
+    boolean retainAll(Collection<?> items);
 
     /**
      * @see java.util.List#clear()