+++ CHANGES.md
@@ -1,12 +1,13 @@
 # Version 3.3.2 (to be released)
 
 * Added convenient constructor to the AbstractComponentDecoration
-* Added WriteOnlyPropertyWrapper to make readable/writable properties write-only
 * Added JTableEditingProperty stating whether a JTable is in the editing state
 * Added JTableRolloverCellProperty holding the position of a table cell under the mouse pointer
 * Added IsParsableRule to check whether a String can be parsed using a Format object
 * Added convenient constructors to the CompositeWritableProperty
-* Added read-only wrappers for set properties, list properties and map properties
+* Added write-ony wrapper for readable/writable properties
+* Added read-only wrappers for readable/writable set properties, list properties and map properties
+* Added write-only wrappers for readable/writable set properties, list properties and map properties
 * Removed inconvenient warning in case the clipping component of a decoration is not showing
 
 # Version 3.3.1